feat(frontend): show member loyalty
